rework Lmod modules for openmpi and hpcx
Rename the openmpi-5.0.8 module to indicate that it only
supports CUDA-based applications on hardware with NVidia GPUs.

Add new wrapper scripts for the NVidia HPCX openmpi libraries.
There are two versions of this - hpcx and hpcx-pmix - that point to
the original library and the one rebuilt with pmix. This provides
conflicts with the other MPI libraries, as well as allows us to
remove the /etc/lmod/.modulespath file that points to the base HPCX
module files.