fix: Improve efficiency and apply overall fixes to the role processes (#22)
* Update info in readme per Dave's suggestions

* Fix dkms builds

* Install kernel-headers and -devel for all installed kernels
* Restart dkms after installing nvidia drivers

* build openmpi with the correct UCX library

openmpi needs to be configured with the rebuilt, PMIx aware UCX
library, not the default on e shipped in HPCX.

Make sure the lmod for the library has the correct library paths set
up, too.

* fix missing libnuma library warning

Testing of the openmpi-5.0.8 library resulted in this warning being
emitted:

[tmp-chhhgt6r-gpu-1:25690] SET UCX_TLS=tcp
[LOG_CAT_SBGP] libnuma.so: cannot open shared object file: No such file
or directory
[LOG_CAT_SBGP] Failed to dlopen libnuma.so. Fallback to GROUP_BY_SOCKET
manual.
[LOG_CAT_SBGP] libnuma.so: cannot open shared object file: No such file
or directory
[LOG_CAT_SBGP] Failed to dlopen libnuma.so. Fallback to GROUP_BY_SOCKET
manual.

The lib64/libnuma.so link is provided by the numactl-devel package,
not numactl-libs. Add the -devel package to the depedency list for
early install.

* rework Lmod modules for openmpi and hpcx

Rename the openmpi-5.0.8 module to indicate that it only
supports CUDA-based applications on hardware with NVidia GPUs.

Add new wrapper scripts for the NVidia HPCX openmpi libraries.
There are two versions of this - hpcx and hpcx-pmix - that point to
the original library and the one rebuilt with pmix. This provides
conflicts with the other MPI libraries, as well as allows us to
remove the /etc/lmod/.modulespath file that points to the base HPCX
module files.